+++ packages/@glimmer/integration-tests/lib/components/emberish-glimmer.ts
@@ -1,5 +1,5 @@
 import { PathReference, UpdatableReference } from '@glimmer/reference';
-import { combine, createUpdatableTag, UpdatableTag, Tag } from '@glimmer/tag';
+import { combine, createUpdatableTag, UpdatableTag, Tag } from '@glimmer/validator';
 import {
   Dict,
   Option,
