+++ packages/@glimmer/integration-tests/lib/modifiers.ts
@@ -8,7 +8,7 @@ import {
   VMArguments,
   CapturedArguments,
 } from '@glimmer/interfaces';
-import { Tag } from '@glimmer/tag';
+import { Tag } from '@glimmer/validator';
 
 export interface TestModifierConstructor {
   new (): TestModifierInstance;
