+++ packages/@glimmer/integration-tests/test/updating-test.ts
@@ -10,7 +10,7 @@ import {
 } from '@glimmer/interfaces';
 import { UpdatableReference } from '@glimmer/object-reference';
 import { ConstReference } from '@glimmer/reference';
-import { bump } from '@glimmer/tag';
+import { bump } from '@glimmer/validator';
 import {
   clientBuilder,
   PrimitiveReference,
