+++ packages/@glimmer/interfaces/lib/components/component-manager.d.ts
@@ -1,5 +1,5 @@
 import { VersionedPathReference } from '@glimmer/reference';
-import { Tag } from '@glimmer/tag';
+import { Tag } from '@glimmer/validator';
 import { SimpleElement } from '@simple-dom/interface';
 import ComponentCapabilities from '../component-capabilities';
 import { ComponentDefinitionState, PreparedArguments, ComponentInstanceState } from '../components';
