+++ packages/@glimmer/interfaces/lib/runtime/modifier.d.ts
@@ -1,7 +1,7 @@
 import { VMArguments } from './arguments';
 import { DynamicScope } from './environment';
 import { GlimmerTreeChanges } from '../dom/changes';
-import { Tag } from '@glimmer/tag';
+import { Tag } from '@glimmer/validator';
 import { Option, SymbolDestroyable, Unique, Destroyable } from '../core';
 import { SimpleElement } from '@simple-dom/interface';
