+++ packages/@glimmer/object-reference/index.ts
@@ -9,4 +9,4 @@ export * from './lib/types';
 export { default as ObjectReference } from './lib/references/path';
 export { default as UpdatableReference, State } from './lib/references/root';
 export { ConstReference } from '@glimmer/reference';
-export { isConst } from '@glimmer/tag';
+export { isConst } from '@glimmer/validator';
