+++ packages/@glimmer/object-reference/lib/references/path.ts
@@ -3,7 +3,7 @@ import { dict } from '@glimmer/util';
 import Meta from '../meta';
 import { PropertyReference } from './descriptors';
 import { PathReference as IPathReference, Reference } from '@glimmer/reference';
-import { Tag, CURRENT_TAG } from '@glimmer/tag';
+import { Tag, CURRENT_TAG } from '@glimmer/validator';
 import { HasGuid } from '@glimmer/util';
 import { Option, Dict } from '@glimmer/interfaces';
