+++ packages/@glimmer/object/lib/object.ts
@@ -6,7 +6,7 @@ const { isArray } = Array;
 
 import { ROOT } from './utils';
 import { Option, Dict } from '@glimmer/interfaces';
-import { bump } from '@glimmer/tag';
+import { bump } from '@glimmer/validator';
 
 export const EMPTY_CACHE = function EMPTY_CACHE() {};
