+++ packages/@glimmer/object/package.json
@@ -6,7 +6,7 @@
     "@glimmer/interfaces": "^0.43.0",
     "@glimmer/object-reference": "^0.43.0",
     "@glimmer/reference": "^0.43.0",
-    "@glimmer/tag": "^0.43.0",
+    "@glimmer/validator": "^0.43.0",
     "@glimmer/util": "^0.43.0"
   },
   "devDependencies": {
