+++ packages/@glimmer/reference/package.json
@@ -5,7 +5,7 @@
   "license": "MIT",
   "dependencies": {
     "@glimmer/util": "^0.43.0",
-    "@glimmer/tag": "^0.43.0"
+    "@glimmer/validator": "^0.43.0"
   },
   "devDependencies": {
     "typescript": "^2.8.3"
