+++ packages/@glimmer/reference/test/iterable-test.ts
@@ -9,7 +9,7 @@ import {
   IteratorSynchronizerDelegate,
   END,
 } from '@glimmer/reference';
-import { Tag, CURRENT_TAG } from '@glimmer/tag';
+import { Tag, CURRENT_TAG } from '@glimmer/validator';
 import { UpdatableReference } from '@glimmer/object-reference';
 
 import { Option, LinkedList, ListNode } from '@glimmer/util';
