+++ packages/@glimmer/runtime/lib/compiled/expressions/concat.ts
@@ -1,7 +1,7 @@
 import { Option, Dict, Maybe } from '@glimmer/interfaces';
 import { CachedReference, PathReference } from '@glimmer/reference';
-import { Tag } from '@glimmer/tag';
-import { combineTagged } from '../../tags';
+import { Tag } from '@glimmer/validator';
+import { combineTagged } from '../../utils/tags';
 
 export class ConcatReference extends CachedReference<Option<string>> {
   public tag: Tag;
