+++ packages/@glimmer/runtime/lib/compiled/opcodes/-debug-strip.ts
@@ -25,7 +25,7 @@ import {
   Option,
 } from '@glimmer/interfaces';
 import { VersionedPathReference, Reference } from '@glimmer/reference';
-import { Tag, COMPUTE } from '@glimmer/tag';
+import { Tag, COMPUTE } from '@glimmer/validator';
 import { ScopeImpl } from '../../environment';
 import CurryComponentReference from '../../references/curry-component';
 import {
