+++ packages/@glimmer/runtime/lib/compiled/opcodes/vm.ts
@@ -1,6 +1,6 @@
 import { PrimitiveType, CompilableTemplate, Option, Op } from '@glimmer/interfaces';
 import { isModified, ReferenceCache } from '@glimmer/reference';
-import { CONSTANT_TAG, isConst, Revision, Tag, value, validate } from '@glimmer/tag';
+import { CONSTANT_TAG, isConst, Revision, Tag, value, validate } from '@glimmer/validator';
 import { initializeGuid, assert } from '@glimmer/util';
 import {
   CheckNumber,
