+++ packages/@glimmer/runtime/lib/opcodes.ts
@@ -3,7 +3,7 @@ import { LowLevelVM, VM, UpdatingVM } from './vm';
 import { Option, Slice as ListSlice, initializeGuid, fillNulls, assert } from '@glimmer/util';
 import { recordStackSize, opcodeMetadata } from '@glimmer/debug';
 import { $pc, $sp, $ra, $fp } from '@glimmer/vm';
-import { Tag } from '@glimmer/tag';
+import { Tag } from '@glimmer/validator';
 import { RuntimeOp, Op, JitOrAotBlock, Maybe, Dict } from '@glimmer/interfaces';
 import { DEBUG, DEVMODE } from '@glimmer/local-debug-flags';
 // these import bindings will be stripped from build
