+++ packages/@glimmer/runtime/lib/references/class-list.ts
@@ -1,8 +1,8 @@
 import { Reference } from '@glimmer/reference';
-import { Tag } from '@glimmer/tag';
+import { Tag } from '@glimmer/validator';
 import { Option } from '@glimmer/util';
 
-import { combineTagged } from '../tags';
+import { combineTagged } from '../utils/tags';
 import { normalizeStringValue } from '../dom/normalize';
 
 export default class ClassListReference implements Reference<Option<string>> {
