+++ packages/@glimmer/runtime/lib/utils/tags.ts
@@ -1,5 +1,13 @@
 import { Slice, LinkedListNode } from '@glimmer/util';
-import { Tag, Tagged, createCombinatorTag, CONSTANT_TAG } from '@glimmer/tag';
+import { Tag, Tagged, createCombinatorTag, CONSTANT_TAG } from '@glimmer/validator';
+
+/**
+ * These utility functions are related to @glimmer/validator, but they aren't
+ * meant to be consumed publicly. They exist as an optimization, and pull in
+ * types that are otherwise unrelated to the validation system. Keeping them
+ * here keeps the validation system isolated, and allows it to avoid pulling in
+ * extra type information (which can lead to issues in public types).
+ */
 
 export function combineTagged(tagged: ReadonlyArray<Tagged>): Tag {
   let optimized: Tag[] = [];
