+++ packages/@glimmer/runtime/package.json
@@ -8,7 +8,7 @@
     "@glimmer/low-level": "^0.43.0",
     "@glimmer/util": "^0.43.0",
     "@glimmer/reference": "^0.43.0",
-    "@glimmer/tag": "^0.43.0",
+    "@glimmer/validator": "^0.43.0",
     "@glimmer/wire-format": "^0.43.0",
     "@glimmer/interfaces": "^0.43.0",
     "@glimmer/program": "^0.43.0",
