+++ packages/@glimmer/validator/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@glimmer/tag",
+  "name": "@glimmer/validator",
   "version": "0.43.0",
   "description": "Objects used to track values and their dirtiness in Glimmer",
   "license": "MIT",
